requirements.txt
torch
transformers
pandas
numpy
seqeval
google-generativeai
tensorboard
scikit-learn
tqdm
matplotlib
seaborn
